refactor(Rule): 오목과 육목의 중복 로직 분리
